Move Mode to parser crate
